requirements.txt
flask>=2.3.0
werkzeug>=2.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
tensorflow>=2.13.0
